Remove reserve from push_list_elements on JIT stack.
Reserve will issue a (non-binding) request to shrink stack size
to the amount you requested to be reserved.  But the whole
point of a stack is that you can extra space if you need to
do nested calls.  Getting rid of the space is a bad idea!
Another failure mode: if we repeatedly push lists of size one,
we'll have quadratic behavior.  Bad.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

ghstack-source-id: 789ec10f9e2a8db8a092b5f3ef6148f7091bcbbc
Pull Request resolved: https://github.com/pytorch/pytorch/pull/27140