Add a test case to RpcTest
# Problem

Existing test cases are too symmetric, so that didn't detect this error, request sent to the wrong worker.

Because of wrong `worker_names` setup, worker0 sends request to itself, while it should had sent to worker1.

# Solution

Add a test case, letting the dst side to check if it's an request from the expected src.

Differential Revision: [D17069062](https://our.internmc.facebook.com/intern/diff/D17069062/)

ghstack-source-id: 91299312
Pull Request resolved: https://github.com/pytorch/pytorch/pull/27322