Fix for flaky caffe2 dataio test (test_time_limit_reader_with_short_limit)
The caffe2 data reader test `test_time_limit_reader_with_short_limit` is flaky as-written because it places an upper bound on how much can be read, but under stress it is possible for fewer records to be read. The fix is to make the assertion check a fuzzy/range check rather than exact equality, since there's not a straightforward way to precisely test a timer-based feature.

Differential Revision: [D17816775](https://our.internmc.facebook.com/intern/diff/D17816775/)

ghstack-source-id: 91543898
Pull Request resolved: https://github.com/pytorch/pytorch/pull/27592