Delete TEST_NAMEDTENSOR; run named tensor tests on all CIs
There's nothing special about the named tensor tests that requires that
they be run in their own CI job. In this PR we delete the
TEST_NAMEDTENSOR flag that hides named tensor tests from regular jobs.
In the future, we'll delete the named tensor CI job so that we do not
duplicate signals.

Test Plan:
- wait for CI

ghstack-source-id: f0715d06781861d89f8f802a862983c4da09fe45
Pull Request resolved: https://github.com/pytorch/pytorch/pull/27760