Fix most documentation warnings
Warnings show up when running `make html` to build documentation. All of
the warnings are very reasonable and point to bugs in our docs. This PR
attempts to fix most of those warnings.

In the future we will add something to the CI that asserts that there
are no warnings in our docs.

Test Plan:
- build and view changes locally

ghstack-source-id: ffb24ded0b17a47e2e11b54ff6e11f47cfbb10c4
Pull Request resolved: https://github.com/pytorch/pytorch/pull/27782