[jit] kill _parameter_list
This was devised in a time when we didn't have module attributes. They
are essentially just tensor lists, so represent them that way. This has
the additional benefit of making the RNN forward pass faster because we
effectively cache the flattened weights.

The only complication part is that someone may come along and do:
```
my_rnn_mod.w_ih_l0 = torch.nn.Parameter(...)
```

This means we need to override setattr to keep the flattened weights
cache up to date.

ghstack-source-id: e8eda6002b9070122df0f24207fd7f2054ea8c06
Pull Request resolved: https://github.com/pytorch/pytorch/pull/27399