[pytorch] Fix pushLong() issue in pickler.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/28057

For pushLong() in Pickler, it looks like we only use for a single use case, with a 10-byte value.

We were handling > 256 bytes incorrectly, by using a LONG4 opcode (expecting 4-byte length), but pushing 8 bytes. We could harden this handling, but rather than improve codepaths that we never expect to use, this change simply removes the incorrect codepath and adds and assert.

ghstack-source-id: 92048325

Differential Revision: [D17934174](https://our.internmc.facebook.com/intern/diff/D17934174/)