Move Method::lowered_graph to a separate pass out of the Method class.
There is no reason to have it in a general API of Module/Method - it's
just another graph pass. It was there because some time ago modules were
not first class and all graphs were lowered. After that changed, this
API was added for easier transition, but now we don't need it anymore.

ghstack-source-id: 499f1a2bbc7e35ed7b5bb7f475c3b6d650d36af1
Pull Request resolved: https://github.com/pytorch/pytorch/pull/28242