Make a dedicated struct for CppHooksList, so we can forward declare it
Continuing the idea of moving AutogradMeta to c10, hooks_list refers to
Variable (which I don't want to move), so I also need to forward declare it.
It's implemented as a shared_ptr (so that we can reflect updates to it
everywhere), so I can also forward declare it fine.

I do wonder if I should keep going this route of forward declaring all of the
things I don't want to move.  By the final count, there's two.  But what
if people add more metadata?  (On the other hand, AutogradMeta changes very
rarely.  So probably we're fine.)

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

ghstack-source-id: b09b18bb6d5b8087929192ef568712d2d8fd98fa
Pull Request resolved: https://github.com/pytorch/pytorch/pull/28240